Make the reporter responsible for saving output files when needed.
